rename: module moscore.py -> moving_window.py
