+++ skchange/change_detectors/moving_window.py
