+++ skchange/change_detectors/tests/test_moscore.py
@@ -3,7 +3,7 @@
 import numpy as np
 import pytest
 
-from skchange.change_detectors.moscore import MovingWindow
+from skchange.change_detectors.moving_window import MovingWindow
 from skchange.change_scores import CHANGE_SCORES
 from skchange.costs import COSTS
 from skchange.datasets.generate import generate_alternating_data
